requirements.txt
langchain>=0.1.0
langchain-community>=0.0.20
langchain-chroma>=0.1.0
duckduckgo-search>=4.0.0
langchainhub>=0.1.0
langchain-google-genai>=1.0.0
InstructorEmbedding>=1.0.0
sentence-transformers==2.2.2
pypdf>=4.0.0
flask>=3.0.0
flask-cors>=4.0.0
pyngrok>=7.0.0
python-dotenv>=1.0.0
chromadb>=0.4.0